+++ src/test/java/network/radicle/jetbrains/radiclejetbrainsplugin/ActionsTest.java
@@ -6,18 +6,11 @@ import com.intellij.dvcs.push.PushSpec;
 import com.intellij.dvcs.push.PushTarget;
 import com.intellij.dvcs.repo.Repository;
 import com.intellij.execution.configurations.GeneralCommandLine;
-import com.intellij.notification.Notification;
 import com.intellij.notification.NotificationGroupManager;
 import com.intellij.notification.NotificationType;
-import com.intellij.notification.Notifications;
 import com.intellij.openapi.actionSystem.AnActionEvent;
-import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.util.SystemInfo;
-import com.intellij.testFramework.HeavyPlatformTestCase;
-import com.intellij.util.messages.MessageBusConnection;
 import com.intellij.vcs.log.VcsFullCommitDetails;
-import git4idea.config.GitConfigUtil;
-import git4idea.repo.GitRepository;
 import network.radicle.jetbrains.radiclejetbrainsplugin.actions.BasicAction;
 import network.radicle.jetbrains.radiclejetbrainsplugin.actions.RadiclePullAction;
 import network.radicle.jetbrains.radiclejetbrainsplugin.actions.RadiclePushAction;
@@ -25,83 +18,20 @@ import network.radicle.jetbrains.radiclejetbrainsplugin.actions.RadicleSyncActio
 import network.radicle.jetbrains.radiclejetbrainsplugin.actions.rad.RadPull;
 import network.radicle.jetbrains.radiclejetbrainsplugin.actions.rad.RadSync;
 import network.radicle.jetbrains.radiclejetbrainsplugin.config.RadicleSettings;
-import network.radicle.jetbrains.radiclejetbrainsplugin.config.RadicleSettingsHandler;
 import network.radicle.jetbrains.radiclejetbrainsplugin.services.RadicleProjectService;
 import org.jetbrains.annotations.NotNull;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.LinkOption;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.TimeUnit;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 @RunWith(JUnit4.class)
-public class ActionsTest extends HeavyPlatformTestCase {
-    private static final Logger logger = Logger.getInstance(ActionsTest.class);
-    public final BlockingQueue<Notification> notificationsQueue = new LinkedBlockingQueue<>();
-    public static final String radVersion = "0.6.0";
-    public static final String radPath = "/usr/bin/rad";
-    public static final String wsl = "wsl";
-    private static final String remoteName = "testRemote";
-    private RadicleSettingsHandler radicleSettingsHandler;
-    private String remoteRepoPath;
-    public GitRepository repository;
-
-    private MessageBusConnection mbc;
-    public RadStub radStub;
-
-    @Before
-    public void before() throws IOException {
-        /* initialize a git repository */
-        remoteRepoPath = Files.createTempDirectory(remoteName).toRealPath(LinkOption.NOFOLLOW_LINKS).toString();
-        repository = GitTestUtil.createGitRepository(super.getProject(),remoteRepoPath);
-
-        /* set rad path */
-        radicleSettingsHandler = new RadicleSettingsHandler();
-        radicleSettingsHandler.loadSettings();
-        radicleSettingsHandler.savePath(radPath);
-
-        /* initialize rad stub service */
-        radStub = RadStub.replaceRadicleApplicationService(this);
-
-        /* add seed node in config */
-        addSeedNodeInConfig();
-        initializeProject();
-
-        mbc = getProject().getMessageBus().connect();
-        mbc.setDefaultHandler(
-                (event1, params) -> {
-                    assertThat(params).hasSize(1);
-                    assertThat(params[0]).isInstanceOf(Notification.class);
-                    Notification n = (Notification) params[0];
-                    logger.warn("captured notification: " + n);
-                    notificationsQueue.add(n);
-                });
-        mbc.subscribe(Notifications.TOPIC);
-        logger.warn("created message bus connection and subscribed to notifications: {}" + mbc);
-    }
-
-    @After
-    public final void after() {
-        if(mbc != null) {
-            mbc.disconnect();
-        }
-        try {
-            repository.dispose();
-        } catch (Exception e) {
-            logger.warn("error disposing git repo");
-        }
-    }
+public class ActionsTest extends AbstractIT {
 
     @Test
     public void testRadPushAction() throws InterruptedException {
@@ -237,17 +167,6 @@ public class ActionsTest extends HeavyPlatformTestCase {
         initializeProject();
     }
 
-    public static void assertCmd(GeneralCommandLine cmd) {
-        assertThat(cmd).isNotNull();
-        if (SystemInfo.isWindows) {
-            assertThat(cmd.getExePath()).isEqualTo(wsl);
-            var path = cmd.getParametersList().get(0);
-            assertThat(path).isEqualTo(radPath);
-        } else {
-            assertThat(cmd.getExePath()).isEqualTo(radPath);
-        }
-    }
-
     public  void assertPushAction() throws InterruptedException {
         var srv = super.getProject().getService(RadicleProjectService.class);
         var cmd = radStub.commands.poll(10, TimeUnit.SECONDS);
@@ -269,39 +188,7 @@ public class ActionsTest extends HeavyPlatformTestCase {
         assertThat(srv.forceRadPush).isFalse();
     }
 
-    private void removeRemoteRadUrl() {
-        try {
-            GitConfigUtil.setValue(super.getProject(),repository.getRoot(), "remote.rad.url","");
-        } catch (Exception e) {
-            logger.warn("unable to write remote rad url in config file");
-        }
-    }
 
-    private void initializeProject() {
-        try {
-            GitConfigUtil.setValue(super.getProject(),repository.getRoot(),
-                    "remote.rad.url","rad://hrrkbjxncopa15doj7qobxip8fotbcemjro4o.git");
-        } catch (Exception e) {
-            logger.warn("unable to write remote rad url in config file");
-        }
-    }
-
-    private void removeSeedNodeFromConfig() {
-        try {
-            GitConfigUtil.setValue(super.getProject(),repository.getRoot(), "rad.seed","");
-        } catch (Exception e) {
-            logger.warn("unable to remove seed node from config file");
-        }
-    }
-
-    private void addSeedNodeInConfig() {
-        try {
-           GitConfigUtil.setValue(super.getProject(),repository.getRoot(), "rad.seed",
-                   "https://maple.radicle.garden");
-        } catch (Exception e) {
-            logger.warn("unable to write seed node in config file");
-        }
-    }
 
     private class PushDetails implements PushInfo {
