+++ src/test/java/network/radicle/jetbrains/radiclejetbrainsplugin/config/RadicleSettingsViewTest.java
@@ -3,6 +3,7 @@ package network.radicle.jetbrains.radiclejetbrainsplugin.config;
 import com.intellij.openapi.options.ConfigurationException;
 import com.intellij.openapi.util.SystemInfo;
 import com.intellij.testFramework.LightPlatform4TestCase;
+import network.radicle.jetbrains.radiclejetbrainsplugin.AbstractIT;
 import network.radicle.jetbrains.radiclejetbrainsplugin.ActionsTest;
 import network.radicle.jetbrains.radiclejetbrainsplugin.RadStub;
 import org.junit.Before;
@@ -71,10 +72,10 @@ public class RadicleSettingsViewTest extends LightPlatform4TestCase {
     public void getRadPathTest() throws InterruptedException {
         var path = radicleSettingsView.getRadPath();
         var cmd = radStub.commands.poll(10, TimeUnit.SECONDS);
-        assertThat(path).isEqualTo(ActionsTest.radPath);
+        assertThat(path).isEqualTo(AbstractIT.radPath);
         assertThat(cmd).isNotNull();
         if (SystemInfo.isWindows) {
-            assertThat(cmd.getExePath()).isEqualTo(ActionsTest.wsl);
+            assertThat(cmd.getExePath()).isEqualTo(AbstractIT.wsl);
             assertThat(cmd.getParametersList().get(0)).isEqualTo(".");
         } else {
             assertThat(cmd.getExePath()).isEqualTo(".");
@@ -84,14 +85,14 @@ public class RadicleSettingsViewTest extends LightPlatform4TestCase {
 
     @Test
     public void getRadVersion() throws InterruptedException {
-        radicleSettingsView.getPathField().setText(ActionsTest.radPath);
+        radicleSettingsView.getPathField().setText(AbstractIT.radPath);
         radicleSettingsView.apply();
 
         var version = radicleSettingsView.getRadVersion();
         var cmd = radStub.commands.poll(10, TimeUnit.SECONDS);
-        ActionsTest.assertCmd(cmd);
-        assertThat(version).isEqualTo(ActionsTest.radVersion);
-        assertThat(cmd.getCommandLineString()).contains(ActionsTest.radPath + " --version");
+        AbstractIT.assertCmd(cmd);
+        assertThat(version).isEqualTo(AbstractIT.radVersion);
+        assertThat(cmd.getCommandLineString()).contains(AbstractIT.radPath + " --version");
     }
 
 }
